Remove deprecated resources
