+++ mongodbatlas/data_source_mongodbatlas_third_party_integrations_test.go
@@ -11,6 +11,7 @@ import (
 )
 
 func TestAccConfigDSThirdPartyIntegrations_basic(t *testing.T) {
+	SkipTest(t)
 	var (
 		projectID       = os.Getenv("MONGODB_ATLAS_PROJECT_ID")
 		intgTypes       = []string{"NEW_RELIC", "OPS_GENIE", "DATADOG", "VICTOR_OPS", "WEBHOOK", "PROMETHEUS"}
