+++ mongodbatlas/provider.go
@@ -162,14 +162,10 @@ func getDataSourcesMap() map[string]*schema.Resource {
 		"mongodbatlas_projects":                                 dataSourceMongoDBAtlasProjects(),
 		"mongodbatlas_cluster":                                  dataSourceMongoDBAtlasCluster(),
 		"mongodbatlas_clusters":                                 dataSourceMongoDBAtlasClusters(),
-		"mongodbatlas_cloud_provider_snapshot":                  dataSourceMongoDBAtlasCloudProviderSnapshot(),
-		"mongodbatlas_cloud_provider_snapshots":                 dataSourceMongoDBAtlasCloudProviderSnapshots(),
 		"mongodbatlas_network_container":                        dataSourceMongoDBAtlasNetworkContainer(),
 		"mongodbatlas_network_containers":                       dataSourceMongoDBAtlasNetworkContainers(),
 		"mongodbatlas_network_peering":                          dataSourceMongoDBAtlasNetworkPeering(),
 		"mongodbatlas_network_peerings":                         dataSourceMongoDBAtlasNetworkPeerings(),
-		"mongodbatlas_cloud_provider_snapshot_restore_job":      dataSourceMongoDBAtlasCloudProviderSnapshotRestoreJob(),
-		"mongodbatlas_cloud_provider_snapshot_restore_jobs":     dataSourceMongoDBAtlasCloudProviderSnapshotRestoreJobs(),
 		"mongodbatlas_maintenance_window":                       dataSourceMongoDBAtlasMaintenanceWindow(),
 		"mongodbatlas_auditing":                                 dataSourceMongoDBAtlasAuditing(),
 		"mongodbatlas_team":                                     dataSourceMongoDBAtlasTeam(),
@@ -185,7 +181,6 @@ func getDataSourcesMap() map[string]*schema.Resource {
 		"mongodbatlas_privatelink_endpoints_service_serverless": dataSourceMongoDBAtlasPrivateLinkEndpointsServiceServerless(),
 		"mongodbatlas_privatelink_endpoint_service_adl":         dataSourceMongoDBAtlasPrivateLinkEndpointServiceADL(),
 		"mongodbatlas_privatelink_endpoints_service_adl":        dataSourceMongoDBAtlasPrivateLinkEndpointsServiceADL(),
-		"mongodbatlas_cloud_provider_snapshot_backup_policy":    dataSourceMongoDBAtlasCloudProviderSnapshotBackupPolicy(),
 		"mongodbatlas_cloud_backup_schedule":                    dataSourceMongoDBAtlasCloudBackupSchedule(),
 		"mongodbatlas_third_party_integrations":                 dataSourceMongoDBAtlasThirdPartyIntegrations(),
 		"mongodbatlas_third_party_integration":                  dataSourceMongoDBAtlasThirdPartyIntegration(),
@@ -236,9 +231,7 @@ func getResourcesMap() map[string]*schema.Resource {
 		"mongodbatlas_database_user":                           resourceMongoDBAtlasDatabaseUser(),
 		"mongodbatlas_project":                                 resourceMongoDBAtlasProject(),
 		"mongodbatlas_cluster":                                 resourceMongoDBAtlasCluster(),
-		"mongodbatlas_cloud_provider_snapshot":                 resourceMongoDBAtlasCloudProviderSnapshot(),
 		"mongodbatlas_network_container":                       resourceMongoDBAtlasNetworkContainer(),
-		"mongodbatlas_cloud_provider_snapshot_restore_job":     resourceMongoDBAtlasCloudProviderSnapshotRestoreJob(),
 		"mongodbatlas_network_peering":                         resourceMongoDBAtlasNetworkPeering(),
 		"mongodbatlas_encryption_at_rest":                      resourceMongoDBAtlasEncryptionAtRest(),
 		"mongodbatlas_private_ip_mode":                         resourceMongoDBAtlasPrivateIPMode(),
@@ -255,7 +248,6 @@ func getResourcesMap() map[string]*schema.Resource {
 		"mongodbatlas_privatelink_endpoint_service":            resourceMongoDBAtlasPrivateEndpointServiceLink(),
 		"mongodbatlas_privatelink_endpoint_service_adl":        resourceMongoDBAtlasPrivateLinkEndpointServiceADL(),
 		"mongodbatlas_privatelink_endpoint_service_serverless": resourceMongoDBAtlasPrivateLinkEndpointServiceServerless(),
-		"mongodbatlas_cloud_provider_snapshot_backup_policy":   resourceMongoDBAtlasCloudProviderSnapshotBackupPolicy(),
 		"mongodbatlas_third_party_integration":                 resourceMongoDBAtlasThirdPartyIntegration(),
 		"mongodbatlas_project_ip_access_list":                  resourceMongoDBAtlasProjectIPAccessList(),
 		"mongodbatlas_cloud_provider_access":                   resourceMongoDBAtlasCloudProviderAccess(),
