--- mongodbatlas/resource_mongodbatlas_cloud_provider_snapshot.go
@@ -1,292 +1,0 @@
-package mongodbatlas
-
-import (
-	"context"
-	"fmt"
-	"log"
-	"net/http"
-	"time"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-
-	matlas "go.mongodb.org/atlas/mongodbatlas"
-)
-
-func resourceMongoDBAtlasCloudProviderSnapshot() *schema.Resource {
-	return &schema.Resource{
-		CreateContext: resourceMongoDBAtlasCloudProviderSnapshotCreate,
-		ReadContext:   resourceMongoDBAtlasCloudProviderSnapshotRead,
-		DeleteContext: resourceMongoDBAtlasCloudProviderSnapshotDelete,
-		Importer: &schema.ResourceImporter{
-			StateContext: resourceMongoDBAtlasCloudProviderSnapshotImportState,
-		},
-		Schema: map[string]*schema.Schema{
-			"project_id": {
-				Type:     schema.TypeString,
-				Required: true,
-				ForceNew: true,
-			},
-			"cluster_name": {
-				Type:     schema.TypeString,
-				Required: true,
-				ForceNew: true,
-			},
-			"description": {
-				Type:     schema.TypeString,
-				Required: true,
-				ForceNew: true,
-			},
-			"retention_in_days": {
-				Type:     schema.TypeInt,
-				Required: true,
-				ForceNew: true,
-			},
-			"created_at": {
-				Type:     schema.TypeString,
-				Computed: true,
-			},
-			"expires_at": {
-				Type:     schema.TypeString,
-				Computed: true,
-			},
-			"master_key_uuid": {
-				Type:     schema.TypeString,
-				Computed: true,
-			},
-			"mongod_version": {
-				Type:     schema.TypeString,
-				Computed: true,
-			},
-			"snapshot_type": {
-				Type:     schema.TypeString,
-				Computed: true,
-			},
-			"status": {
-				Type:     schema.TypeString,
-				Computed: true,
-			},
-			"storage_size_bytes": {
-				Type:     schema.TypeInt,
-				Computed: true,
-			},
-			"type": {
-				Type:     schema.TypeString,
-				Computed: true,
-			},
-			"snapshot_id": {
-				Type:     schema.TypeString,
-				Computed: true,
-			},
-			"timeout": {
-				Type:     schema.TypeString,
-				Optional: true,
-				ForceNew: true,
-				Default:  "10m",
-			},
-		},
-		DeprecationMessage: "this resource is deprecated, and will be removed in v1.9 release, please transition as soon as possible to mongodbatlas_cloud_backup_snapshot",
-	}
-}
-
-func resourceMongoDBAtlasCloudProviderSnapshotRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
-	// Get client connection.
-	conn := meta.(*MongoDBClient).Atlas
-	ids := decodeStateID(d.Id())
-
-	requestParameters := &matlas.SnapshotReqPathParameters{
-		SnapshotID:  ids["snapshot_id"],
-		GroupID:     ids["project_id"],
-		ClusterName: ids["cluster_name"],
-	}
-
-	snapshotReq, resp, err := conn.CloudProviderSnapshots.GetOneCloudProviderSnapshot(context.Background(), requestParameters)
-	if err != nil {
-		if resp != nil && resp.StatusCode == http.StatusNotFound {
-			d.SetId("")
-			return nil
-		}
-
-		return diag.FromErr(fmt.Errorf("error getting snapshot Information: %s", err))
-	}
-
-	if err = d.Set("snapshot_id", snapshotReq.ID); err != nil {
-		return diag.FromErr(fmt.Errorf("error setting `snapshot_id` for snapshot (%s): %s", ids["snapshot_id"], err))
-	}
-
-	if err = d.Set("created_at", snapshotReq.CreatedAt); err != nil {
-		return diag.FromErr(fmt.Errorf("error setting `created_at` for snapshot (%s): %s", ids["snapshot_id"], err))
-	}
-
-	if err = d.Set("expires_at", snapshotReq.ExpiresAt); err != nil {
-		return diag.FromErr(fmt.Errorf("error setting `expires_at` for snapshot (%s): %s", ids["snapshot_id"], err))
-	}
-
-	if err = d.Set("master_key_uuid", snapshotReq.MasterKeyUUID); err != nil {
-		return diag.FromErr(fmt.Errorf("error setting `master_key_uuid` for snapshot (%s): %s", ids["snapshot_id"], err))
-	}
-
-	if err = d.Set("mongod_version", snapshotReq.MongodVersion); err != nil {
-		return diag.FromErr(fmt.Errorf("error setting `mongod_version` for snapshot (%s): %s", ids["snapshot_id"], err))
-	}
-
-	if err = d.Set("snapshot_type", snapshotReq.SnapshotType); err != nil {
-		return diag.FromErr(fmt.Errorf("error setting `snapshot_type` for snapshot (%s): %s", ids["snapshot_id"], err))
-	}
-
-	if err = d.Set("status", snapshotReq.Status); err != nil {
-		return diag.FromErr(fmt.Errorf("error setting `status` for snapshot (%s): %s", ids["snapshot_id"], err))
-	}
-
-	if err = d.Set("storage_size_bytes", snapshotReq.StorageSizeBytes); err != nil {
-		return diag.FromErr(fmt.Errorf("error setting `storage_size_bytes` for snapshot (%s): %s", ids["snapshot_id"], err))
-	}
-
-	if err = d.Set("type", snapshotReq.Type); err != nil {
-		return diag.FromErr(fmt.Errorf("error setting `type` for snapshot (%s): %s", ids["snapshot_id"], err))
-	}
-
-	return nil
-}
-
-func resourceMongoDBAtlasCloudProviderSnapshotCreate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
-	// Get client connection.
-	conn := meta.(*MongoDBClient).Atlas
-
-	requestParameters := &matlas.SnapshotReqPathParameters{
-		GroupID:     d.Get("project_id").(string),
-		ClusterName: d.Get("cluster_name").(string),
-	}
-
-	snapshotReq := &matlas.CloudProviderSnapshot{
-		Description:     d.Get("description").(string),
-		RetentionInDays: d.Get("retention_in_days").(int),
-	}
-
-	var timeout time.Duration
-
-	timeout, err := time.ParseDuration(d.Get("timeout").(string))
-	if err != nil {
-		return diag.FromErr(err)
-	}
-
-	stateConf := &resource.StateChangeConf{
-		Pending:    []string{"CREATING", "UPDATING", "REPAIRING", "REPEATING"},
-		Target:     []string{"IDLE"},
-		Refresh:    resourceClusterRefreshFunc(ctx, d.Get("cluster_name").(string), d.Get("project_id").(string), conn),
-		Timeout:    timeout,
-		MinTimeout: 10 * time.Second,
-		Delay:      3 * time.Minute,
-	}
-
-	// Wait, catching any errors
-	_, err = stateConf.WaitForStateContext(ctx)
-	if err != nil {
-		return diag.FromErr(err)
-	}
-
-	snapshot, _, err := conn.CloudProviderSnapshots.Create(ctx, requestParameters, snapshotReq)
-	if err != nil {
-		return diag.FromErr(fmt.Errorf("error taking a snapshot: %s", err))
-	}
-
-	requestParameters.SnapshotID = snapshot.ID
-
-	stateConf = &resource.StateChangeConf{
-		Pending:    []string{"queued", "inProgress"},
-		Target:     []string{"completed", "failed"},
-		Refresh:    resourceCloudProviderSnapshotRefreshFunc(ctx, requestParameters, conn),
-		Timeout:    1 * time.Hour,
-		MinTimeout: 60 * time.Second,
-		Delay:      1 * time.Minute,
-	}
-
-	// Wait, catching any errors
-	_, err = stateConf.WaitForStateContext(ctx)
-	if err != nil {
-		return diag.FromErr(err)
-	}
-
-	d.SetId(encodeStateID(map[string]string{
-		"project_id":   d.Get("project_id").(string),
-		"cluster_name": d.Get("cluster_name").(string),
-		"snapshot_id":  snapshot.ID,
-	}))
-
-	return resourceMongoDBAtlasCloudProviderSnapshotRead(ctx, d, meta)
-}
-
-func resourceMongoDBAtlasCloudProviderSnapshotDelete(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
-	// Get client connection.
-	conn := meta.(*MongoDBClient).Atlas
-	ids := decodeStateID(d.Id())
-
-	requestParameters := &matlas.SnapshotReqPathParameters{
-		SnapshotID:  ids["snapshot_id"],
-		GroupID:     ids["project_id"],
-		ClusterName: ids["cluster_name"],
-	}
-
-	_, err := conn.CloudProviderSnapshots.Delete(ctx, requestParameters)
-	if err != nil {
-		return diag.FromErr(fmt.Errorf("error deleting a snapshot (%s): %s", ids["snapshot_id"], err))
-	}
-
-	return nil
-}
-
-func resourceCloudProviderSnapshotRefreshFunc(ctx context.Context, requestParameters *matlas.SnapshotReqPathParameters, client *matlas.Client) resource.StateRefreshFunc {
-	return func() (interface{}, string, error) {
-		c, resp, err := client.CloudProviderSnapshots.GetOneCloudProviderSnapshot(ctx, requestParameters)
-
-		switch {
-		case err != nil:
-			return nil, "failed", err
-		case resp.StatusCode == http.StatusNotFound:
-			return "", "DELETED", nil
-		case c.Status == "failed":
-			return nil, c.Status, fmt.Errorf("error creating MongoDB snapshot(%s) status was: %s", requestParameters.SnapshotID, c.Status)
-		}
-
-		if c.Status != "" {
-			log.Printf("[DEBUG] status for MongoDB snapshot: %s: %s", requestParameters.SnapshotID, c.Status)
-		}
-
-		return c, c.Status, nil
-	}
-}
-
-func resourceMongoDBAtlasCloudProviderSnapshotImportState(ctx context.Context, d *schema.ResourceData, meta interface{}) ([]*schema.ResourceData, error) {
-	conn := meta.(*MongoDBClient).Atlas
-
-	requestParameters, err := splitSnapshotImportID(d.Id())
-	if err != nil {
-		return nil, err
-	}
-
-	u, _, err := conn.CloudProviderSnapshots.GetOneCloudProviderSnapshot(ctx, requestParameters)
-	if err != nil {
-		return nil, fmt.Errorf("couldn't import snapshot %s in project %s, error: %s", requestParameters.ClusterName, requestParameters.GroupID, err)
-	}
-
-	d.SetId(encodeStateID(map[string]string{
-		"project_id":   requestParameters.GroupID,
-		"cluster_name": requestParameters.ClusterName,
-		"snapshot_id":  requestParameters.SnapshotID,
-	}))
-
-	if err := d.Set("project_id", requestParameters.GroupID); err != nil {
-		log.Printf("[WARN] Error setting project_id for (%s): %s", requestParameters.SnapshotID, err)
-	}
-
-	if err := d.Set("cluster_name", requestParameters.ClusterName); err != nil {
-		log.Printf("[WARN] Error setting cluster_name for (%s): %s", requestParameters.SnapshotID, err)
-	}
-
-	if err := d.Set("description", u.Description); err != nil {
-		log.Printf("[WARN] Error setting description for (%s): %s", requestParameters.SnapshotID, err)
-	}
-
-	return []*schema.ResourceData{d}, nil
-}
