+++ mongodbatlas/resource_mongodbatlas_project_ip_access_list_test.go
@@ -154,7 +154,7 @@ func TestAccProjectRSProjectIPAccessList_SettingMultiple(t *testing.T) {
 	resourceName := "mongodbatlas_project_ip_access_list.test_%d"
 	orgID := os.Getenv("MONGODB_ATLAS_ORG_ID")
 	projectName := acctest.RandomWithPrefix("test-acc")
-	const ipWhiteListCount = 200
+	const ipWhiteListCount = 20
 	accessList := make([]map[string]string, 0)
 
 	for i := 0; i < ipWhiteListCount; i++ {
